refactor/ride: define services interfaces and extract implementation
